src/requirement.txt
absl-py==1.1.0
apturl==0.5.2
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
autocolorize==0.2.1
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==5.2.0
caffeine==2.9.8
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
dill==0.3.5.1
distro==1.7.0
distro-info===1.1build1
Django==4.0.5
djangorestframework==3.13.1
duplicity==0.8.21
etils==0.6.0
ewmh==0.1.6
executing==0.8.3
fasteners==0.14.1
flatbuffers==1.12
fonttools==4.33.3
future==0.18.2
gast==0.4.0
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.3
gpg===1.16.0-unknown
grpcio==1.46.3
h5py==3.7.0
httplib2==0.20.2
idna==3.3
imageio==2.19.2
importlib-metadata==4.6.4
importlib-resources==5.8.0
ipython==8.4.0
jedi==0.18.1
jeepney==0.7.1
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.4.3
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
networkx==2.8.2
numpy==1.22.4
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.9.0
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.12.0
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyswarms==1.3.0
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-xlib==0.29
pytz==2022.1
PyWavelets==1.3.0
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
screen-resolution-extra==0.0.0
seaborn==0.11.2
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
systemd-python==234
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.9.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.5.4
toml==0.10.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.2.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
